+++ src/routes/auth.py
@@ -1,3 +1,13 @@
 from flask import Blueprint
 
 auth = Blueprint('auth', __name__, url_prefix='/auth')
+
+
+@auth.route('/login')
+def login():
+    pass
+
+
+@auth.route('/logout')
+def logout():
+    pass